+++ Code/Engine/Core/Platform/Linux/InputDevice_Linux.cpp
@@ -1,0 +1,5 @@
+#include <Core/CorePCH.h>
+
+#if EZ_ENABLED(EZ_PLATFORM_LINUX)
+#  include <Core/Platform/NoImpl/InputDevice_NoImpl.h>
+#endif
