+++ Code/Engine/Core/Platform/OSX/InputDevice_OSX.cpp
@@ -1,0 +1,5 @@
+#include <Core/CorePCH.h>
+
+#if EZ_ENABLED(EZ_PLATFORM_OSX)
+#  include <Core/Platform/NoImpl/InputDevice_NoImpl.h>
+#endif
