+++ Code/Engine/Core/Platform/OSX/Window_OSX.cpp
@@ -1,0 +1,5 @@
+#include <Core/CorePCH.h>
+
+#if EZ_ENABLED(EZ_PLATFORM_OSX)
+#  include <Core/Platform/NoImpl/Window_NoImpl.h>
+#endif
