+++ Code/Engine/Core/System/Implementation/Window.cpp
@@ -8,26 +8,6 @@
 #include <Foundation/IO/OpenDdlWriter.h>
 #include <Foundation/System/Screen.h>
 
-#if EZ_ENABLED(EZ_SUPPORTS_GLFW)
-#  include <Core/System/Implementation/glfw/InputDevice_glfw.inl>
-#  include <Core/System/Implementation/glfw/Window_glfw.inl>
-#elif EZ_ENABLED(EZ_PLATFORM_WINDOWS_DESKTOP)
-#  include <Core/System/Implementation/Win/InputDevice_win32.inl>
-#  include <Core/System/Implementation/Win/Window_win32.inl>
-#elif EZ_ENABLED(EZ_PLATFORM_WINDOWS_UWP)
-#  include <Core/System/Implementation/uwp/InputDevice_uwp.inl>
-#  include <Core/System/Implementation/uwp/Window_uwp.inl>
-#elif EZ_ENABLED(EZ_PLATFORM_ANDROID)
-#  include <Core/System/Implementation/android/InputDevice_android.inl>
-#  include <Core/System/Implementation/android/Window_android.inl>
-#elif EZ_ENABLED(EZ_PLATFORM_WEB)
-#  include <Core/System/Implementation/Web/InputDevice_Web.h>
-#  include <Core/System/Implementation/Web/Window_Web.inl>
-#else
-#  include <Core/System/Implementation/null/InputDevice_null.inl>
-#  include <Core/System/Implementation/null/Window_null.inl>
-#endif
-
 ezUInt8 ezWindow::s_uiNextUnusedWindowNumber = 0;
 
 ezResult ezWindowCreationDesc::AdjustWindowSizeAndPosition()
